+++ spec/unit/generators/templates/turnip_spec.rb
@@ -18,19 +18,20 @@ RSpec.describe Howitzer::BaseGenerator do
 
             AllCops:
               DisplayCopNames: true
-              TargetRubyVersion: 2.4
+              NewCops: enable
+              TargetRubyVersion: 3
 
             Layout/CaseIndentation:
               Enabled: false
+
+            Layout/LineLength:
+              Max: 120
 
             Lint/AmbiguousRegexpLiteral:
               Enabled: false
 
             Metrics/BlockLength:
               Enabled: false
-
-            Metrics/LineLength:
-              Max: 120
 
             Metrics/ModuleLength:
               Max: 150